refactor(#303): 공통 컴포넌트 분리
