+++ src/app/(auth)/_components/SignIn/AuthButtons/GoogleAuth.tsx
@@ -1,6 +1,7 @@
 import { useGoogleLogin } from '@react-oauth/google';
 import { GoogleSVG } from '@/icons/svg';
 
+/* eslint-disable no-unused-vars */
 interface GoogleAuthProps {
   onSuccess: (token: string) => void;
   onError: (response: any) => void;
@@ -13,6 +14,7 @@ declare global {
   }
 }
 
+/* eslint-enable no-unused-vars */
 const GoogleAuth = ({ onSuccess, onError }: GoogleAuthProps) => {
   const login = useGoogleLogin({
     onSuccess: (tokenResponse) => onSuccess(tokenResponse.access_token),
